improve tests for timerange extractor
