+++ nsdb-cluster/src/test/resources/application.conf
@@ -69,7 +69,7 @@ nsdb {
   publisher.timeout = 30 seconds
   publisher.scheduler.interval = 5 seconds
 
-  write.scheduler.interval = 15 seconds
+  write.scheduler.interval = 5 seconds
 
   stream.timeout = 30 seconds
 }
