+++ nsdb-cluster/src/test/scala/io/radicalbit/nsdb/cluster/coordinator/ReadCoordinatorSimpleStatementsSpec.scala
@@ -16,193 +16,16 @@
 
 package io.radicalbit.nsdb.cluster.coordinator
 
-import akka.actor.{ActorSystem, Props}
-import akka.pattern.ask
-import akka.testkit.{ImplicitSender, TestKit, TestProbe}
-import akka.util.Timeout
-import com.typesafe.config.{ConfigFactory, ConfigValueFactory}
-import io.radicalbit.nsdb.cluster.actor.MetricsDataActor
-import io.radicalbit.nsdb.cluster.actor.MetricsDataActor.AddRecordToLocation
-import io.radicalbit.nsdb.cluster.coordinator.MetadataCoordinator.commands.AddLocation
-import io.radicalbit.nsdb.cluster.coordinator.mockedActors.{LocalMetadataCache, LocalMetadataCoordinator}
 import io.radicalbit.nsdb.common.protocol._
 import io.radicalbit.nsdb.common.statement._
 import io.radicalbit.nsdb.index.{BIGINT, DECIMAL, VARCHAR}
-import io.radicalbit.nsdb.model.{Location, SchemaField}
+import io.radicalbit.nsdb.model.SchemaField
 import io.radicalbit.nsdb.protocol.MessageProtocol.Commands._
 import io.radicalbit.nsdb.protocol.MessageProtocol.Events._
-import org.scalatest._
 
-import scala.concurrent.Await
 import scala.concurrent.duration._
 
-object LongMetric {
-
-  val name = "longMetric"
-
-  val recordsShard1: Seq[Bit] = Seq(
-    Bit(1L, 1L, Map("surname" -> "Doe"), Map("name" -> "John")),
-    Bit(2L, 2L, Map("surname" -> "Doe"), Map("name" -> "John")),
-    Bit(4L, 3L, Map("surname" -> "D"), Map("name"   -> "J"))
-  )
-
-  val recordsShard2: Seq[Bit] = Seq(
-    Bit(6L, 4L, Map("surname"  -> "Doe"), Map("name" -> "Bill")),
-    Bit(8L, 5L, Map("surname"  -> "Doe"), Map("name" -> "Frank")),
-    Bit(10L, 6L, Map("surname" -> "Doe"), Map("name" -> "Frankie"))
-  )
-
-  val testRecords = recordsShard1 ++ recordsShard2
-}
-
-object DoubleMetric {
-
-  val name = "doubleMetric"
-
-  val recordsShard1: Seq[Bit] = Seq(
-    Bit(2L, 1.5, Map("surname" -> "Doe"), Map("name" -> "John")),
-    Bit(4L, 1.5, Map("surname" -> "Doe"), Map("name" -> "John"))
-  )
-
-  val recordsShard2: Seq[Bit] = Seq(
-    Bit(6L, 1.5, Map("surname"  -> "Doe"), Map("name" -> "Bill")),
-    Bit(8L, 1.5, Map("surname"  -> "Doe"), Map("name" -> "Frank")),
-    Bit(10L, 1.5, Map("surname" -> "Doe"), Map("name" -> "Frankie"))
-  )
-
-  val testRecords = recordsShard1 ++ recordsShard2
-}
-
-object AggregationMetric {
-
-  val name = "aggregationMetric"
-
-  val recordsShard1: Seq[Bit] = Seq(
-    Bit(2L, 2L, Map("surname" -> "Doe"), Map("name" -> "John", "age" -> 15L, "height" -> 30.5)),
-    Bit(4L, 2L, Map("surname" -> "Doe"), Map("name" -> "John", "age" -> 20L, "height" -> 30.5))
-  )
-
-  val recordsShard2: Seq[Bit] = Seq(
-    Bit(2L, 1L, Map("surname"  -> "Doe"), Map("name" -> "John", "age"    -> 15L, "height" -> 30.5)),
-    Bit(6L, 1L, Map("surname"  -> "Doe"), Map("name" -> "Bill", "age"    -> 15L, "height" -> 31.0)),
-    Bit(8L, 1L, Map("surname"  -> "Doe"), Map("name" -> "Frank", "age"   -> 15L, "height" -> 32.0)),
-    Bit(10L, 1L, Map("surname" -> "Doe"), Map("name" -> "Frankie", "age" -> 15L, "height" -> 32.0))
-  )
-
-  val testRecords = recordsShard1 ++ recordsShard2
-}
-
-class ReadCoordinatorSpec
-    extends TestKit(
-      ActorSystem(
-        "ReadCoordinatorSpec",
-        ConfigFactory
-          .load()
-          .withValue("nsdb.sharding.interval", ConfigValueFactory.fromAnyRef("5s"))
-      ))
-    with ImplicitSender
-    with WordSpecLike
-    with Matchers
-    with BeforeAndAfterAll
-    with WriteInterval {
-
-  val probe     = TestProbe()
-  val basePath  = "target/test_index/ReadCoordinatorShardSpec"
-  val db        = "db"
-  val namespace = "registry"
-  val schemaCoordinator =
-    system.actorOf(SchemaCoordinator.props(basePath, system.actorOf(Props[FakeSchemaCache])), "schemacoordinator")
-  val metadataCoordinator =
-    system.actorOf(LocalMetadataCoordinator.props(system.actorOf(Props[LocalMetadataCache])), "metadatacoordinator")
-  val writeCoordinator =
-    system.actorOf(WriteCoordinator.props(metadataCoordinator, schemaCoordinator, system.actorOf(Props.empty)))
-  val metricsDataActor =
-    system.actorOf(MetricsDataActor.props(basePath, "node1", writeCoordinator))
-  val readCoordinatorActor = system actorOf ReadCoordinator.props(metadataCoordinator,
-                                                                  schemaCoordinator,
-                                                                  system.actorOf(Props.empty))
-
-  override def beforeAll = {
-    import scala.concurrent.duration._
-    implicit val timeout = Timeout(5.second)
-
-    Await.result(readCoordinatorActor ? SubscribeMetricsDataActor(metricsDataActor, "node1"), 10 seconds)
-
-    val location1 = Location(_: String, "node1", 0, 5)
-    val location2 = Location(_: String, "node1", 6, 10)
-
-    //long metric
-    Await.result(
-      metricsDataActor ? DropMetricWithLocations(db,
-                                                 namespace,
-                                                 LongMetric.name,
-                                                 Seq(location1(LongMetric.name), location2(LongMetric.name))),
-      10 seconds)
-
-    Await.result(
-      schemaCoordinator ? UpdateSchemaFromRecord(db, namespace, LongMetric.name, LongMetric.testRecords.head),
-      10 seconds)
-
-    LongMetric.recordsShard1.foreach(r => {
-      Await.result(metadataCoordinator ? AddLocation(db, namespace, location1(LongMetric.name)), 10 seconds)
-      Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location1(LongMetric.name)), 10 seconds)
-    })
-    LongMetric.recordsShard2.foreach(r => {
-      Await.result(metadataCoordinator ? AddLocation(db, namespace, location2(LongMetric.name)), 10 seconds)
-      Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location2(LongMetric.name)), 10 seconds)
-    })
-
-    //double metric
-    Await.result(
-      metricsDataActor ? DropMetricWithLocations(db,
-                                                 namespace,
-                                                 DoubleMetric.name,
-                                                 Seq(location1(DoubleMetric.name), location2(DoubleMetric.name))),
-      10 seconds)
-
-    Await.result(
-      schemaCoordinator ? UpdateSchemaFromRecord(db, namespace, DoubleMetric.name, DoubleMetric.testRecords.head),
-      10 seconds)
-
-    DoubleMetric.recordsShard1.foreach(r => {
-      Await.result(metadataCoordinator ? AddLocation(db, namespace, location1(DoubleMetric.name)), 10 seconds)
-      Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location1(DoubleMetric.name)), 10 seconds)
-    })
-    DoubleMetric.recordsShard2.foreach(r => {
-      Await.result(metadataCoordinator ? AddLocation(db, namespace, location2(DoubleMetric.name)), 10 seconds)
-      Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location2(DoubleMetric.name)), 10 seconds)
-    })
-
-    //aggregation metric
-    Await.result(
-      metricsDataActor ? DropMetricWithLocations(db,
-                                                 namespace,
-                                                 AggregationMetric.name,
-                                                 Seq(location1(AggregationMetric.name),
-                                                     location2(AggregationMetric.name))),
-      10 seconds
-    )
-
-    Await.result(schemaCoordinator ? UpdateSchemaFromRecord(db,
-                                                            namespace,
-                                                            AggregationMetric.name,
-                                                            AggregationMetric.testRecords.head),
-                 10 seconds)
-
-    AggregationMetric.recordsShard1.foreach(r => {
-      Await.result(metadataCoordinator ? AddLocation(db, namespace, location1(AggregationMetric.name)), 10 seconds)
-      Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location1(AggregationMetric.name)),
-                   10 seconds)
-    })
-    AggregationMetric.recordsShard2.foreach(r => {
-      Await.result(metadataCoordinator ? AddLocation(db, namespace, location2(AggregationMetric.name)), 10 seconds)
-      Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location2(AggregationMetric.name)),
-                   10 seconds)
-    })
-
-    expectNoMessage(interval)
-    expectNoMessage(interval)
-  }
+class ReadCoordinatorSimpleStatementsSpec extends AbstractReadCoordinatorSpec {
 
   "ReadCoordinator" when {
 
@@ -936,365 +759,6 @@ class ReadCoordinatorSpec
         }
         expected.values.size should be(1)
       }
-    }
-
-    "receive a select containing a GTE selection and a group by" should {
-      "execute it successfully" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = LongMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(SumAggregation)))),
-              condition = Some(Condition(
-                ComparisonExpression(dimension = "timestamp", comparison = GreaterOrEqualToOperator, value = 2L))),
-              groupBy = Some(SimpleGroupByAggregation("name"))
-            )
-          )
-        )
-
-        val expected = awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }
-        expected.values.size should be(5)
-      }
-    }
-
-    "receive a select containing a GTE selection and a group by without any aggregation" should {
-      "fail" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = LongMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("creationDate", None))),
-              condition = Some(Condition(
-                ComparisonExpression(dimension = "timestamp", comparison = GreaterOrEqualToOperator, value = 2L))),
-              groupBy = Some(SimpleGroupByAggregation("name"))
-            )
-          )
-        )
-        awaitAssert {
-          probe.expectMsgType[SelectStatementFailed]
-        }
-      }
-    }
-
-    "receive a select containing a non existing entity" should {
-      "return an error message properly" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(db = db,
-                               namespace = namespace,
-                               metric = "nonexisting",
-                               distinct = false,
-                               fields = AllFields,
-                               limit = Some(LimitOperator(5)))
-          )
-        )
-        awaitAssert {
-          probe.expectMsgType[SelectStatementFailed]
-        }
-      }
-    }
-
-    "receive a select containing a group by on string dimension " should {
-      "execute it successfully when count(*) is used instead of value" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = LongMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("*", Some(CountAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("name")),
-              order = Some(AscOrderOperator("name"))
-            )
-          )
-        )
-
-        val expected = awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }
-        expected.values shouldBe Seq(
-          Bit(0L, 1L, Map.empty, Map("name" -> "Bill")),
-          Bit(0L, 1L, Map.empty, Map("name" -> "Frank")),
-          Bit(0L, 1L, Map.empty, Map("name" -> "Frankie")),
-          Bit(0L, 1L, Map.empty, Map("name" -> "J")),
-          Bit(0L, 2L, Map.empty, Map("name" -> "John"))
-        )
-      }
-
-      "execute it successfully with asc ordering over string dimension" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = LongMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(CountAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("name")),
-              order = Some(AscOrderOperator("name"))
-            )
-          )
-        )
-
-        val expected = awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }
-        expected.values shouldBe Seq(
-          Bit(0L, 1L, Map.empty, Map("name" -> "Bill")),
-          Bit(0L, 1L, Map.empty, Map("name" -> "Frank")),
-          Bit(0L, 1L, Map.empty, Map("name" -> "Frankie")),
-          Bit(0L, 1L, Map.empty, Map("name" -> "J")),
-          Bit(0L, 2L, Map.empty, Map("name" -> "John"))
-        )
-      }
-
-      "execute it successfully with desc ordering over string dimension" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = LongMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(SumAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("name")),
-              order = Some(DescOrderOperator("name"))
-            )
-          )
-        )
-
-        awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }.values shouldBe Seq(
-          Bit(0L, 3, Map.empty, Map("name" -> "John")),
-          Bit(0L, 3, Map.empty, Map("name" -> "J")),
-          Bit(0L, 6, Map.empty, Map("name" -> "Frankie")),
-          Bit(0L, 5, Map.empty, Map("name" -> "Frank")),
-          Bit(0L, 4, Map.empty, Map("name" -> "Bill"))
-        )
-
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = DoubleMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(SumAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("name")),
-              order = Some(DescOrderOperator("name"))
-            )
-          )
-        )
-
-        awaitAssert {
-          val expected = probe.expectMsgType[SelectStatementExecuted]
-          expected.values shouldBe Seq(
-            Bit(0L, 3.0, Map.empty, Map("name" -> "John")),
-            Bit(0L, 1.5, Map.empty, Map("name" -> "Frankie")),
-            Bit(0L, 1.5, Map.empty, Map("name" -> "Frank")),
-            Bit(0L, 1.5, Map.empty, Map("name" -> "Bill"))
-          )
-        }
-      }
-
-      "execute it successfully with desc ordering over numerical dimension" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = LongMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(SumAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("name")),
-              order = Some(DescOrderOperator("value"))
-            )
-          )
-        )
-
-        awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }.values.map(_.value) shouldBe Seq(6, 5, 4, 3, 3)
-
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = DoubleMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(SumAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("name")),
-              order = Some(DescOrderOperator("value"))
-            )
-          )
-        )
-
-        awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }.values.map(_.value) shouldBe Seq(3.0, 1.5, 1.5, 1.5)
-
-      }
-
-      "execute it successfully with asc ordering over numerical dimension" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = LongMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(SumAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("name")),
-              order = Some(AscOrderOperator("value")),
-              limit = Some(LimitOperator(2))
-            )
-          )
-        )
-
-        awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }.values.map(_.value) shouldBe Seq(3, 3)
-
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = DoubleMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(SumAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("name")),
-              order = Some(AscOrderOperator("value")),
-              limit = Some(LimitOperator(2))
-            )
-          )
-        )
-
-        awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }.values.map(_.value) shouldBe Seq(1.5, 1.5)
-      }
-    }
-
-    "receive a select containing a group by on long dimension" should {
-      "execute it successfully with count aggregation" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = AggregationMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(CountAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("age")),
-              order = Some(AscOrderOperator("value"))
-            )
-          )
-        )
-
-        awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }.values shouldBe Seq(Bit(0L, 1, Map.empty, Map("age" -> 20)), Bit(0L, 5, Map.empty, Map("age" -> 15)))
-      }
-
-      "execute it successfully with sum aggregation" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = AggregationMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(SumAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("age")),
-              order = Some(AscOrderOperator("age"))
-            )
-          )
-        )
-
-        awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }.values shouldBe Seq(
-          Bit(0L, 6L, Map.empty, Map("age" -> 15L)),
-          Bit(0L, 2L, Map.empty, Map("age" -> 20L))
-        )
-      }
-    }
-
-    "receive a select containing a group by on double dimension" should {
-      "execute it successfully with count aggregation" in within(5.seconds) {
-        probe.send(
-          readCoordinatorActor,
-          ExecuteStatement(
-            SelectSQLStatement(
-              db = db,
-              namespace = namespace,
-              metric = AggregationMetric.name,
-              distinct = false,
-              fields = ListFields(List(Field("value", Some(CountAggregation)))),
-              groupBy = Some(SimpleGroupByAggregation("height")),
-              order = Some(DescOrderOperator("value"))
-            )
-          )
-        )
-
-        awaitAssert {
-          probe.expectMsgType[SelectStatementExecuted]
-        }.values shouldBe Seq(
-          Bit(0L, 3, Map.empty, Map("height" -> 30.5)),
-          Bit(0L, 2, Map.empty, Map("height" -> 32.0)),
-          Bit(0L, 1, Map.empty, Map("height" -> 31.0))
-        )
-      }
-    }
-
-    "execute it successfully with sum aggregation" in within(5.seconds) {
-      probe.send(
-        readCoordinatorActor,
-        ExecuteStatement(
-          SelectSQLStatement(
-            db = db,
-            namespace = namespace,
-            metric = AggregationMetric.name,
-            distinct = false,
-            fields = ListFields(List(Field("value", Some(SumAggregation)))),
-            groupBy = Some(SimpleGroupByAggregation("height")),
-            order = Some(AscOrderOperator("height"))
-          )
-        )
-      )
-
-      awaitAssert {
-        probe.expectMsgType[SelectStatementExecuted]
-      }.values shouldBe Seq(
-        Bit(0L, 5, Map.empty, Map("height" -> 30.5)),
-        Bit(0L, 1, Map.empty, Map("height" -> 31.0)),
-        Bit(0L, 2, Map.empty, Map("height" -> 32.0))
-      )
     }
   }
 }
