+++ nsdb-common/src/main/scala/io/radicalbit/nsdb/common/statement/SQLStatement.scala
@@ -151,7 +151,16 @@ sealed trait GroupByAggregation {
   def dimension: String
 }
 
+/**
+  * Class that represent a simple Group By clause.
+  * @param dimension the dimension to apply the aggregation to
+  */
 case class SimpleGroupByAggregation(dimension: String) extends GroupByAggregation
+
+/**
+  * Temporal aggregation.
+  * @param interval The time aggregation interval in Milliseconds.
+  */
 case class TemporalGroupByAggregation(interval: Long) extends GroupByAggregation {
   override val dimension: String = "timestamp"
 }
