+++ nsdb-core/src/test/scala/io/radicalbit/nsdb/statement/StatementParserSpec.scala
@@ -1004,84 +1004,5 @@ class StatementParserSpec extends WordSpec with Matchers {
       }
 
     }
-
-    "executing computeRangeForInterval " should {
-      "return a seq of ranges" in {
-        val res = StatementParser.computeRangeForInterval(10L, 0L, 5L, Seq.empty)
-        res.size shouldBe 2
-        res.map(_.label).contains("5-10") shouldBe true
-        res.map(_.label).contains("0-5") shouldBe true
-      }
-    }
-
-    "executing computeRanges " should {
-      "return a seq of ranges for a RangeExpression" in {
-        val res = StatementParser.computeRanges(
-          5L,
-          None,
-          Some(Condition(RangeExpression(dimension = "timestamp", value1 = 0L, value2 = 10L))),
-          100L)
-        res.size shouldBe 2
-        res.map(_.label).contains("5-10") shouldBe true
-        res.map(_.label).contains("0-5") shouldBe true
-      }
-      "return a seq of ranges for a left bounded interval(>=) lower than now " in {
-        val res = StatementParser.computeRanges(
-          5L,
-          None,
-          Some(
-            Condition(
-              ComparisonExpression(dimension = "timestamp", comparison = GreaterOrEqualToOperator, value = 10L))),
-          100L)
-        res.size shouldBe 10
-      }
-      "return a seq of ranges for a right bounded interval(<=) lower than now " in {
-        val res = StatementParser.computeRanges(
-          5L,
-          None,
-          Some(
-            Condition(ComparisonExpression(dimension = "timestamp", comparison = LessOrEqualToOperator, value = 100L))),
-          200L)
-
-        res.exists(r => r.max == 99 && r.min == 95) shouldBe true
-        res.size shouldBe 10
-      }
-      "return a seq of ranges for both right and left bounded interval( >= && <=) " in {
-        val res = StatementParser.computeRanges(
-          5L,
-          None,
-          Some(
-            Condition(TupledLogicalExpression(
-              ComparisonExpression(dimension = "timestamp", comparison = GreaterOrEqualToOperator, value = 70L),
-              AndOperator,
-              ComparisonExpression(dimension = "timestamp", comparison = LessOrEqualToOperator, value = 100L)
-            ))),
-          200L
-        )
-
-        res.exists(r => r.max == 99 && r.min == 95) shouldBe true
-        res.exists(r => r.max == 74 && r.min == 70) shouldBe true
-        res.size shouldBe 6
-      }
-
-      "return a seq of ranges for both right and left bounded interval( > && <) " in {
-        val res = StatementParser.computeRanges(
-          5L,
-          None,
-          Some(
-            Condition(TupledLogicalExpression(
-              ComparisonExpression(dimension = "timestamp", comparison = GreaterThanOperator, value = 70L),
-              AndOperator,
-              ComparisonExpression(dimension = "timestamp", comparison = LessThanOperator, value = 100L)
-            ))),
-          200L
-        )
-
-        res.exists(r => r.max == 94 && r.min == 90) shouldBe true
-        res.exists(r => r.max == 74 && r.min == 70) shouldBe true
-        res.size shouldBe 5
-      }
-
-    }
   }
 }
